+++ src/pages/CSS/basicPage.css
@@ -40,10 +40,6 @@
     /* color: black; */
 }
 
-.question span {
-    /* border: 1px solid grey; */
-}
-
 .QuestionNum {
     font-size: 22px;
     margin-left: 6px;
