+++ src/pages/CSS/detailedPage.css
