+++ src/pages/HTML/DetailedPage.tsx
@@ -3,8 +3,8 @@ import { Form, ProgressBar, Alert,
 Offcanvas, OffcanvasHeader,OffcanvasTitle, Row, Col, Container} from "react-bootstrap";
 import OpenAI from "openai";
 import { key } from "./homePage"
-import "./detailedPage.css";
-import LoaderComp from "../components/loader";
+import "../CSS/detailedPage.css";
+import LoaderComp from "../../components/loader";
 import { parseAnswers } from "./ResultsPage";
 
 const QUESTIONSTARTS = //first half of the questions to make response formatting for GPT easier, will add the user's response to the end of the strings
